study-session/pre-active-learn/requirements.txt
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
google-cloud-storage>=2.10.0
google-genai>=0.2.0
google-generativeai>=0.3.0
google-adk
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
PyPDF2>=3.0.0
